Done until the forklidt phase of cutting wood.

+++ mtgvrp/job_manager/lumberjack/TreeItem.cs
@@ -13,6 +13,7 @@ using MongoDB.Bson;
 using GTANetworkShared;
 using mtgvrp.database_manager;
 using mtgvrp.door_manager;
+using mtgvrp.player_manager;
 using MongoDB.Bson.Serialization.Attributes;
 using MongoDB.Driver;
 
@@ -57,6 +58,21 @@ namespace mtgvrp.job_manager.lumberjack
         [BsonIgnore]
         public int CutPercentage { get; set; }
 
+        [BsonIgnore]
+        public int ProcessPercentage { get; set; }
+
+        public enum Stages
+        {
+            Cutting,
+            Processing,
+            Waiting,
+            Moving,
+            Hidden,
+        }
+
+        [BsonIgnore]
+        public Stages Stage = Stages.Cutting;
+
         public void Insert()
         {
             DatabaseManager.TreesTable.InsertOne(this);
@@ -84,26 +100,52 @@ namespace mtgvrp.job_manager.lumberjack
 
         public void CreateTree()
         {
-            TreeObj = API.shared.createObject(-1279773008, TreePos, TreeRot);
-            TreeText = API.shared.createTextLabel("~g~" + CutPercentage + "% Cut.~n~Tree", TreePos.Add(new Vector3(1, 0, 1)), 5f, 1f, true);
+            UpdateAllTree();
         }
 
         public void UpdateTreeText()
         {
-            if (API.shared.doesEntityExist(TreeText))
+            if (TreeText != null && API.shared.doesEntityExist(TreeText))
                 API.shared.deleteEntity(TreeText);
 
-            TreeText = API.shared.createTextLabel("~g~" + CutPercentage + "% Cut.~n~Tree", TreePos.Add(new Vector3(1, 0, 1)), 5f, 1f, true);
+            switch (Stage)
+            {
+                case Stages.Processing:
+                    TreeText = API.shared.createTextLabel("~g~Hit the tree to process it. ~n~" + ProcessPercentage + "%", TreePos.Add(new Vector3(0.5, -1, 1)), 5f, 1f, true);
+                    break;
+                case Stages.Cutting:
+                    TreeText = API.shared.createTextLabel("~g~" + CutPercentage + "% Cut.~n~Tree", TreePos.Add(new Vector3(1, 0, 1)), 5f, 1f, true);
+                    break;
+                case Stages.Waiting:
+                    TreeText = API.shared.createTextLabel("~g~Waiting to be picked, use /pickupwood with a Forklift.", TreePos.Add(new Vector3(1, 0, 1)), 5f, 1f, true);
+                    break;
+
+                default:
+                    TreeText = null;
+                    break;
+            }
+        }
+
+        public void UpdateTreeObject()
+        {
+            if (TreeObj != null && API.shared.doesEntityExist(TreeObj))
+                API.shared.deleteEntity(TreeObj);
+
+            switch (Stage)
+            {
+                case Stages.Waiting:
+                    TreeObj = API.shared.createObject(-1186441238, TreePos, TreeRot);
+                    break;
+                default:
+                    TreeObj = API.shared.createObject(-1279773008, TreePos, TreeRot);
+                    break;
+            }
         }
 
         public void UpdateAllTree()
         {
             UpdateTreeText();
-
-            if (API.shared.doesEntityExist(TreeObj))
-                API.shared.deleteEntity(TreeObj);
-
-            TreeObj = API.shared.createObject(-1279773008, TreePos, TreeRot);
+            UpdateTreeObject();
         }
 
         public static void LoadTrees()
